feat(project): Add call of readline with prompt. Separate libs tests in tests.c. Create tests.h for tests functions
