+++ include/tests.h
@@ -1,0 +1,8 @@
+#ifndef TESTS_H
+# define TESTS_H
+
+void	test_gnl(void);
+void	test_printf(void);
+void	test_libft(void);
+
+#endif
